requirements.txt
attrs==22.1.0
bcrypt==4.0.1
blinker==1.5
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coverage==6.5.0
exceptiongroup==1.0.1
Faker==15.3.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
flask-expects-json==1.7.0
Flask-Mail==0.9.1
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
packaging==21.3
pluggy==1.0.0
protobuf==3.20.1
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.12
Werkzeug==2.2.2
